\name{spExpM-package}
\alias{spExpM-package}
\alias{spExpM}
\docType{package}
\title{
\packageTitle{spExpM}
}
\description{
\packageDescription{spExpM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{spExpM}
\packageIndices{spExpM}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{spExpM}

Maintainer: \packageMaintainer{spExpM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
